🔥 feat(seat_number_widget): remove unused imports and code
The changes made to the `seat_number_widget.dart` file include the removal of unused imports and commented-out code. This helps to improve the code's readability and maintainability by reducing the amount of unnecessary code.

The main changes are:

1. Removed the `AwesomeDialog` and `MyFlashBar` imports and their corresponding usages.
2. Commented out the `GetBuilder<CoversSheetsController>` section, which included a dialog for deleting an exam cover sheet.

These changes were made to simplify the `SeatNumberWidget` by removing functionality that is not currently being used, making the code more focused and easier to understand.