Refactor API call to fetch subjects by school type
Previously, the API call to fetch subjects was hardcoded to use the `SubjectsLinks.subjects` path. This commit modifies the API call to dynamically fetch subjects based on the school type stored in the Hive box. The new path is constructed using `${SubjectsLinks.subjectsBySchoolType}/${Hive.box('School').get('SchoolTypeID')}`. This change ensures that the correct set of subjects is fetched based on the user's selected school type.