🔧 refactor: remove unused constants from dio_factory
The changes in this commit remove the unused `CONTENT_TYPE` and `APPLICATION_JSON` constants from the `dio_factory.dart` file. These constants were previously used to set the headers in the `DioFactory` class, but they are no longer needed since the headers are now dynamically generated based on the presence of a token.