🧹 feat(firebase): update hosting configuration
Update the hosting configuration in the `.firebaserc` file to use the
`staging-control-system-2025` name instead of `control-system-2025`. This
change ensures that the staging environment is properly configured for
deployment.

Additionally, update the `AppLinks` class in `lib/app/configurations/app_links.dart`
to include the staging URL for the control system application. This change
provides the correct URL for the staging environment, allowing the application
to correctly access the staging server.