🚀 feat(class_room_res_model): add class_number field
The changes add a new field `classNumber` to the `ClassRoomResModel` class. This field represents the class number and is used to store the class number information. The changes also update the `toJson()` and `fromJson()` methods to handle the new field.

These changes are important as they allow the application to store and retrieve the class number information for each classroom.