Add leading space to class number text in ClassRoomsWidget
In the `ClassRoomsWidget` widget, a leading space was added to the class number text to improve readability. This change ensures that the class number is visually separated from the preceding text.

Additionally, the version number in `pubspec.yaml` was updated from `1.0.0` to `1.0.1` to reflect the changes made in this commit.