Add Tooltips to IconButtons for Improved User Experience
This commit introduces tooltips to various IconButtons across multiple widgets in the application. The addition of tooltips provides users with clear and concise descriptions of the actions associated with each button, enhancing the overall user experience. This change is particularly beneficial for users who may not be familiar with the application's interface or who require additional context for certain actions.

The tooltips were added to the following widgets:
- AdminWidget
- AllUserWidget
- UserInSchoolWidget
- ClassRoomsWidget
- EditClassRoomWidget
- AddSubjectsToCohort
- CohortsWidget
- AddNewStudentsToControlMission
- CreateMissionWidget
- DistributeStudentsSideMenue
- MissionDetailsWidget
- RenderStudentsInEaxmRoom
- CustomAppBar
- FrontIdScreen
- AddNewProctor
- EditProctorWidget
- ExamRoomsAssignedToProctorWidget
- ProctorsInExamRoomWidget
- AddSingleStudentWidget
- EditStudentWidget
- StudentWidget
- TransferStudentWidget
- EditOperationWidget
- EditSubjectsWidget
- SubjectsWidget
- SystemLoggerScreen

Each tooltip is designed to succinctly describe the action that will be performed when the associated IconButton is pressed, ensuring that users can navigate the application with greater confidence and ease.