🐛 fix: disable worker manager logging
This commit disables the worker manager logging by commenting out the `workerManager.log = true` line in the `main.dart` file. This change is made to improve the performance and reduce the noise in the logs.

✨ feat: update app version to 1.7.1

This commit updates the app version in the `pubspec.yaml` file from 1.7.0 to 1.7.1. This change is made to reflect the latest version of the app.

🧹 refactor: improve subject check in review control mission controller

This commit refactors the subject check logic in the `review_control_mission_controller.dart` file. It changes the check from comparing the subject name to comparing the subject ID. This change is made to ensure the correct subject is being checked, as the subject name may not be unique.