Disable User Actions in System Logger Table
Removed the 'ActionsField' column and associated functionality from the system logger table to temporarily disable user actions. This includes commenting out the column definition, the renderer for the action button, and the related user info widget. This change is intended to prevent user actions while a more permanent solution is developed.